+++ source/themes/simple/index.js
@@ -8,6 +8,7 @@ import SimpleBubble from './SimpleBubble.scss';
 import SimpleButton from './SimpleButton.scss';
 import SimpleCheckbox from './SimpleCheckbox.scss';
 import SimpleFormField from './SimpleFormField.scss';
+import SimpleHeader from './SimpleHeader.scss';
 import SimpleInput from './SimpleInput.scss';
 import SimpleLoadingSpinner from './SimpleLoadingSpinner.scss';
 import SimpleModal from './SimpleModal.scss';
@@ -30,6 +31,7 @@ export const SimpleTheme = {
   [IDENTIFIERS.BUTTON]: SimpleButton,
   [IDENTIFIERS.CHECKBOX]: SimpleCheckbox,
   [IDENTIFIERS.FORM_FIELD]: SimpleFormField,
+  [IDENTIFIERS.HEADER]: SimpleHeader,
   [IDENTIFIERS.INPUT]: SimpleInput,
   [IDENTIFIERS.LOADING_SPINNER]: SimpleLoadingSpinner,
   [IDENTIFIERS.MODAL]: SimpleModal,